Create a new class for the class Select method
